+++ example/crao/config-overrides.prod.js
@@ -1,0 +1,3 @@
+
+// config-overrides.prod
+module.exports = (config) => config;