+++ example/src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux";
+import { rtdReducer } from "rt-design";
+
+const reducer = combineReducers({
+  rtd: rtdReducer,
+});
+
+export default reducer;